GeoAI/requirements.txt
fastapi
uvicorn
python-dotenv
@google/generative-ai
jinja2
